requirements.txt
fastapi
uvicorn
pytest
httpx
requests
grpcio
grpcio-tools
flask